+++ simulator/simulator.h
@@ -53,10 +53,10 @@ public:
 
     virtual size_t sizeof_register() const = 0;
 
-    virtual uint64 read_cpu_register( uint8 regno) const = 0;
-    virtual uint64 read_gdb_register( uint8 regno) const = 0;
-    virtual void write_cpu_register( uint8 regno, uint64 value) = 0;
-    virtual void write_gdb_register( uint8 regno, uint64 value) = 0;
+    virtual uint64 read_cpu_register( size_t regno) const = 0;
+    virtual uint64 read_gdb_register( size_t regno) const = 0;
+    virtual void write_cpu_register( size_t regno, uint64 value) = 0;
+    virtual void write_gdb_register( size_t regno, uint64 value) = 0;
     virtual void write_csr_register( std::string_view name, uint64 value) = 0;
 
     int get_exit_code() const { return exit_code; }
